Flyttet loggekode ut av Serviceklasse.
